backend/requirements.txt
# FastAPI
fastapi>=0.128.5
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LangChain
langchain>=1.2.9
langchain-openai>=1.1.7
langchain-anthropic>=1.3.2
langchain-community>=0.4.1
langgraph>=1.0.8

# Database
psycopg2-binary==2.9.10
pgvector==0.3.6
sqlalchemy==2.0.36

# Document Processing
anthropic>=0.49.0
markitdown[xlsx]>=0.1.0
PyMuPDF>=1.25.0
python-docx>=1.1.0

# Authentication
python-jose[cryptography]>=3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
email-validator==2.3.0

# BOX API
box-sdk-gen[jwt]>=1.17.0

# Load Testing
locust>=2.32.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings>=2.10.1
aiofiles==24.1.0
slowapi>=0.1.9
